dbt_duckdb_project/dag_duckdb_scheduler/pyproject.toml
[project]
name = "dag_duckdb_scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "dagster",
    "dagster-cloud",
    "dagster-dbt",
    "dbt-core<1.11",
    "dbt-duckdb<1.11",
    "dbt-duckdb<1.11",
    "dbt-duckdb<1.11",
    "dbt-duckdb<1.11",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "dag_duckdb_scheduler.definitions"
code_location_name = "dag_duckdb_scheduler"